PlotlyLightLite.PlotlyLightLite — Module

PlotlyLightLite is a lightweight interface to the underlying PlotlyLight package (Fastest time-to-first-plot in Julia!), itself an "ultra-lightweight interface" to the Plotly javascript libraries. PlotlyLightLite supports some of the Plots.jl interface for plotting. This package may be of use as an alternative to Plots in resource-constrained environments, such as binder.

Plots.jl uses

positional arguments for data

keyword arguments for attributes

plot() is a workhorse with seriestype indicating which plot; there are also special methods (e.g. scatter(x,y) becomes plot(x,y; seriestype="scatter"))

PlotlyLight.jl uses

data::Vector{Config} to hold tracts of data for plotting

layout::Config to adjust layout

config::Config to adjust global configurations

Config very flexibly creates the underlying Javascript objects the plotly interface expects

Plot() is a workhorse with type acting like seriestype and also mode

PlotlyLightLite.jl has this dispatch for plot:

merge layout; merge config; pass kwargs to Config push onto data or merge onto last tract:

plot(; layout::Config?, config::Config, kwargs...)
plot!([p::Plot]; layout::Config?, config::Config?, kwargs...)

Line plot. connecting x,y (and possibly z). For 2D, use !isfinite values in y to break.

plot(x,y,[z]; kwargs...)
plot!([p::Plot], x, y, [z]; kwargs...

Data can be generated from a function:

plot(f::Function, ab; kwargs...) => plot(unzip(f, ab)...; kwargs...)
plot(f::Function, a, b; kwargs...) => plot(unzip(f, a, b)...; kwargs...)
plot!([p::Plot], f, ab, [b])

plot each function as lineplot:

plot(fs::Vector{Function}, a, [b]; kwargs...)
plot!([p::Plot], fs::Vector{Function}, a, [b]; kwargs...)

Note

Currently x, y make vectors; should matrices be supported using column vectors? "In Plots.jl, every column is a series, a set of related points which form lines, surfaces, or other plotting primitives. "

Parametric line plots, 2 or 3d

plot(fs::NTuple(N,Function), a, [b]; kwargs...)
plot!([p::Plot], fs::NTuple(N,Function), a, [b]; kwargs...)

Alternatively

plot(u::Function, v::Function, [w::Function], a, [b]; kwargs...)

In addition there are these plot constructors for higher-dimensional plots

contour

implicit_plot

heatmap

surface

wireframe

There are also numerous functions to modify attributes of an existing plot.

PlotlyLightLite.annotate! — Method

annotate!([p::Plot], x, y, txt; [color], [family], [pointsize], [halign], [valign])
annotate!([p::Plot], anns::Tuple;  kwargs...)

Add annotations to plot.

x, y, txt: text to add at (x,y)

color: text color

family: font family

pointsize: text size

halign: one of "top", "bottom"

valign: one of "left", "right"

rotation: angle to rotate

The x, y, txt values can be specified as 3 iterables or tuple of tuples.

PlotlyLightLite.arrow — Method

arrow(tails, vs; kwargs...)

Draw vectors vs anchored at tails. Hacked in support for 3D using combination of lines + cones.

Use Plotly attributes arrowcolor, arrowwidth,

PlotlyLightLite.circle! — Method

circle([p::Plot], x0, x1, y0, y1; kwargs...)

Draw circle shape bounded in [x0, x1] × [y0,y1]. (Will adjust to non-equal sized boundary.)

Example

Use named tuple for line for boundary.

circle!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

PlotlyLightLite.contour — Method

contour(x, y, z; kwargs...)
contour!([p::Plot], x, y, z; kwargs...)
contour(x, y, f::Function; kwargs...)
contour!(x, y, f::Function; kwargs...)

Create contour function of f

PlotlyLightLite.current — Method

current()

Get current figure. A Plot object of PlotlyLight; UndefRefError if none.

Not typically needed, as it is implicit in most mutating calls, though may be convenient if those happen within a loop.

PlotlyLightLite.grid_layout — Method

grid_layout(ps::Array{<:Plot})

Layout an array of plots into a grid. Vectors become rows of plots.

Use Plot() to create an empty plot for a given cell.

Example

using DataFrames
n = 25
d = DataFrame(x=sin.(rand(n)), y=rand(n).^2, z = rand(n)) # assume all numeric
nms = names(d)
m = Matrix{Plot}(undef, length(nms), length(nms))

for i ∈ eachindex(nms)
    for j ∈ eachindex(nms)
        if j > i
            p = Plot()
        elseif j == i
            x = d[:,j]
            p = plot(x, type="histogram")
            xlabel!(p, nms[j])  # <<- not working in grid! as differe xaxis purposes
        else
            x = d[:,i]; y = d[:,j]
            p = scatter(x, y)
            xlabel!(p, nms[i]); ylabel!(p, nms[j])
        end
        m[i,j] = p
    end
end
grid_layout(m)

PlotlyLightLite.heatmap — Method

heatmap(x, y, z; kwargs...)
heatmap!([p::Plot], x, y, z; kwargs...)
heatmap(x, y, f::Function; kwargs...)
heatmap!(x, y, f::Function; kwargs...)

Create heatmap function of f

PlotlyLightLite.implicit_plot — Method

implicit_plot(f; xlims=(-5,5), ylims=(-5,5), legend=false, linewidth=2, kwargs...)
implicit_plot!([p::Plot], f; kwargs...)

For f(x,y) = ... plot implicitly defined y(x) from f(x,y(x)) = 0 over range specified by xlims and ylims.

Example

f(x,y) = x * y - (x^3 + x^2 + x + 1)
implicit_plot(f)

(Basically just contour plot with levels=0 and points detremined by extrema of xlims and ylims.)

PlotlyLightLite.legend! — Function

legend!([p::Plot], legend::Bool) hide/show legend

PlotlyLightLite.plot! — Method

plot!([p::Plot], x, y; kwargs...)
plot!([p::Plot], f, a, [b]; kwargs...)
plot!([p::Plot], f; kwargs...)

Used to add a new tract to an existing plot. Like Plots.plot!. See plot for argument details.

PlotlyLightLite.plot — Method

plot(x, y, [z]; [linecolor], [linewidth], [legend], kwargs...)
plot(f::Function, a, [b]; kwargs...)

Create a line plot.

Returns a Plot instance from PlotlyLight

x,y points to plot. NaN values in y break the line

a, b: the interval to plot a function over can be given by two numbers or if just a then by extrema(a).

linecolor: color of line

linewidth: width of line

label in legend

Other keyword arguments include width and height, xlims and ylims, legend, aspect_ratio.

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotlyLight allows full access to the underlying JavaScript library library.

The provided "Plots-like" functions are plot, plot!, scatter!, scatter, annotate!, title!, xlims! and ylims!.

Example

p = plot(sin, 0, 2pi; legend=false)
plot!(cos)
# add points
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
# add text
annotate!(tuple(zip(x0, sin.(x0), ("A", "B"))...), halign="left", pointsize=12)
title!("Sine and cosine and where they intersect in [0,2π]")
# adjust limits
ylims!((-3/2, 3/2))
# add shape
y0, y1 = extrema(p).y
[rect!(xᵢ-0.1, xᵢ+0.1, y0, y1, fillcolor="gray", opacity=0.2) for xᵢ ∈ x0]
# display plot
p

Warning

You may need to run the first plot cell twice to see an image.

PlotlyLightLite.plot — Method

plot(; layout::Config?, config::Config?, kwargs...)

Pass keyword arguments through Config and onto PlotlyLight.Plot.

PlotlyLightLite.plot — Method

plot((f,g), a, b; kwargs...)
plot!([p::PLot], (f,g), a, b; kwargs...)

Make parametric plot from tuple of functions, f and g.

PlotlyLightLite.quiver! — Function

quiver!([p::Plot], x, y, txt=nothing; quiver=(dx, dy), kwargs...)
quiver(x, y, txt=nothing; quiver=(dx, dy), kwargs...)

Draw 2d arrows. See arrow! for a single arrow.

(x,y) are tail positions, optionally labeled by txt

quiver specifies vector part of arrow

kwargs processarrowhead::Int?,arrowwidth::Int?,arrowcolor`

Example

ts = range(0, 2pi, length=100)
p = plot(sin.(ts), cos.(ts), linecolor="red")
ts = range(0, 2pi, length=10)
quiver!(p, cos.(ts), sin.(ts), quiver=(-sin.(ts), cos.(ts)), arrowcolor="red")
p

This example shows how text can be rotated with angles in degrees and positive angles measured in a clockwise direction.

ts = range(0, 2pi, 100)
p = plot(cos.(ts), sin.(ts), linecolor="red", aspect_ratio=:equal,
    linewidth=20, opacity=0.2)

txt = split("The quick brown fox jumped over the lazy dog")
ts = range(0, 360, length(txt)+1)[2:end]
for (i,t) ∈ enumerate(reverse(ts))
    quiver!(p, [cosd(t)],[sind(t)],txt[i],
            quiver=([0],[0]),
            textangle=90-t,
            font=(size=20,))
end
xaxis!(zeroline=false); yaxis!(zeroline=false) # remove zerolines
p

3d arrows

3d arrows are possible using arrows!.

PlotlyLightLite.rect! — Method

rect!([p::Plot], x0, x1, y0, y1; kwargs...)

Draw rectangle shape.

Example

Use named tuple for line for boundary.

rect!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

PlotlyLightLite.scatter! — Method

scatter(x, y, [z]; [markershape], [markercolor], [markersize], kwargs...)
scatter!([p::Plot], x, y, [z]; kwargs...)

Place point on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

PlotlyLightLite.scatter — Method

scatter(x, y; kwargs...) see scatter!

PlotlyLightLite.scroll_zoom! — Method

scrollzoom!([p], x::Bool) turn on/off scrolling to zoom

PlotlyLightLite.size! — Method

size!([p::Plot]; [width], [height]) specify size of plot figure

PlotlyLightLite.surface — Method

surface(x, y, z; kwargs...)
surface!(x, y, z; kwargs...)
surface(x, y, f::Function; kwargs...)
surface!(x, y, f::Function; kwargs...)

Create surface plot. Pass zcontour=true to add contour plot projected onto the z axis.

Example

f(x,y) = 4 - x^2 - y^2
xs = ys = range(-2, 2, length=50)
surface(xs, ys, f)  # zs = f.(xs', ys)

Extended help

From https://discourse.julialang.org/t/3d-surfaces-time-slider/109673

z1 = Vector[[8.83, 8.89, 8.81, 8.87, 8.9, 8.87],
                       [8.89, 8.94, 8.85, 8.94, 8.96, 8.92],
                       [8.84, 8.9, 8.82, 8.92, 8.93, 8.91],
                       [8.79, 8.85, 8.79, 8.9, 8.94, 8.92],
                       [8.79, 8.88, 8.81, 8.9, 8.95, 8.92],
                       [8.8, 8.82, 8.78, 8.91, 8.94, 8.92],
                       [8.75, 8.78, 8.77, 8.91, 8.95, 8.92],
                       [8.8, 8.8, 8.77, 8.91, 8.95, 8.94],
                       [8.74, 8.81, 8.76, 8.93, 8.98, 8.99],
                       [8.89, 8.99, 8.92, 9.1, 9.13, 9.11],
                       [8.97, 8.97, 8.91, 9.09, 9.11, 9.11],
                       [9.04, 9.08, 9.05, 9.25, 9.28, 9.27],
                       [9, 9.01, 9, 9.2, 9.23, 9.2],
                       [8.99, 8.99, 8.98, 9.18, 9.2, 9.19],
                       [8.93, 8.97, 8.97, 9.18, 9.2, 9.18]]
xs , ys = 1:length(z1[1]), 1:length(z1) # needed here given interface chosen
surface(xs, ys, z1, colorscale="Viridis")
surface!(xs, ys, map(x -> x .+ 1, z1), colorscale="Viridis", showscale=false, opacity=0.9)
surface!(xs, ys, map(x -> x .- 1, z1), colorscale="Viridis", showscale=false, opacity=0.9)

Julia users would typically use a matrix to hold the z data, but Javascript users would expect a vector of vectors, as above. As PlotlyLight just passes on the data to Javascript, the above is perfectly acceptable. (The keyword arguments above come from Plotly, not Plots.)

A parameterized surface can be displayed. Below the unzip function returns 3 matrices specifying the surface described by the vector-valued function r.

r1, r2 = 2, 1/2
r(u,v) = ((r1 + r2*cos(v))*cos(u), (r1 + r2*cos(v))*sin(u), r2*sin(v))
us = vs = range(0, 2pi, length=25)
xs, ys, zs = unzip(us, vs, r)

surface(xs, ys, zs)

Plotting a vertical plane poses a slight challenge, as we can't parameterize as z=f(x,y). Here we intersect a surface with the plane ax + by + 0z = d and add a trace for the intersection of the two surfaces.

f(x, y) = 4 - (x^2 + y^2)

# surface
xs =  ys = range(-2, 2, length=100)
zs = f.(xs', ys)

# (vertical) plane
m, M = extrema(zs)
zzs = range(m, M, length=2)
Xs = ((x,z) -> x).(xs', zzs)
Zs = ((x,z) -> z).(xs', zzs)
a, b, c, d = 1, 1, 0, 1
plane(x,z) = (d - a*x - c*z) / b
Ys = plane.(xs', zzs)

# intersection
g(t) = (d - a*t) / b
γ(t) = (t, y(t), f(t, y(t)))

surface(xs, ys, zs)
surface!(Xs, Ys, Zs, opacity=0.2)
plot!(unzip(γ.(xs))...; linewidth=3)

PlotlyLightLite.title! — Method

title!([p::Plot], txt)

Set plot title.

PlotlyLightLite.unzip — Method

unzip(v, [vs...])
unzip(f::Function, a, b)
unzip(a, b, F::Function)

Reshape data to x,y,[z] mode.

In its basic use, zip takes two vectors, pairs them off, and returns an iterator of tuples for each pair. For unzip a vector of same-length vectors is "unzipped" to return two (or more) vectors.

The function version applies f to a range of points over (a,b) and then calls unzip. This uses the adapted_grid function from PlotUtils.

The function version with F computes F(a', b) and then unzips. This is used with parameterized surface plots

This uses the invert function of SplitApplyCombine.

PlotlyLightLite.wireframe — Method

wireframe(x, y, z; kwargs...)
wireframe(x, y, f::Function; kwargs...)
wireframe!([p::Plot], x, y, z; kwargs...)
wireframe!([p::Plot], x, y, f::Function; kwargs...)

Create wireframe.

Example

f(x, y) = 4 - x^2 - y^2
xs = ys = range(-2, 2, length=100)
surface(xs, ys, f)
wireframe!(xs, ys, f)

PlotlyLightLite.xaxis! — Method

xticks!([p::Plot]; kwargs...)
yticks!([p::Plot]; kwargs...)

Adjust ticks on chart.

ticks: a container or range

ticklabels: optional labels (same length as ticks)

showticklabels::Bool

PlotlyLightLite.xlims! — Method

xlims!(p, lims) set x limits of plot

PlotlyLightLite.ylims! — Method

ylims!(p, lims) set y limits of plot