PlotlyLightLite.annotate! — Method

annotate!([p::Plot], x, y, txt; [color], [family], [pointsize], [halign], [valign])
annotate!([p::Plot], anns::Tuple;  kwargs...)

Add annotations to plot.

x, y, txt: text to add at (x,y)

color: text color

family: font family

pointsize: text size

halign: one of "top", "bottom"

valign: one of "left", "right"

rotation: angle to rotate

The x, y, txt values can be specified as 3 iterables or tuple of tuples.

PlotlyLightLite.circle! — Method

circle([p::Plot], x0, x1, y0, y1; kwargs...)

Draw circle shape bounded in [x0, x1] × [y0,y1]. (Will adjust to non-equal sized boundary.)

Example

Use named tuple for line for boundary.

circle!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

PlotlyLightLite.contour — Method

countour(x, y, z; kwargs...)
contour!([p::Plot], x, y, z; kwargs...)
contour(x, y, f::Function; kwargs...)
contour!(x, y, f::Function; kwargs...)

Create contour function of f

PlotlyLightLite.current — Method

current()

Get current figure. A Plot object of PlotlyLight; UndefRefError if none.

Not typically needed, as it is implicit in most mutating calls, though may be convenient if those happen within a loop.

PlotlyLightLite.grid_layout — Method

grid_layout(ps::Array{<:Plot})

Layout an array of plots into a grid. Vectors become rows of plots.

Use Plot() to create an empty plot for a given cell.

Example

using DataFrames
n = 25
d = DataFrame(x=sin.(rand(n)), y=rand(n).^2, z = rand(n)) # assume all numeric
nms = names(d)
m = Matrix{Plot}(undef, length(nms), length(nms))

for i ∈ eachindex(nms)
    for j ∈ eachindex(nms)
        if j > i
            p = Plot()
        elseif j == i
            x = d[:,j]
            p = plot(x, type="histogram")
            xlabel!(p, nms[j])  # <<- not working in grid! as differe xaxis purposes
        else
            x = d[:,i]; y = d[:,j]
            p = scatter(x, y)
            xlabel!(p, nms[i]); ylabel!(p, nms[j])
        end
        m[i,j] = p
    end
end
grid_layout(m)

PlotlyLightLite.legend! — Function

legend!([p::Plot], legend::Bool) hide/show legend

PlotlyLightLite.plot — Function

plot((f,g), a, b; kwargs...)
plot!([p::PLot], (f,g), a, b; kwargs...)

Make parametric plot from tuple of functions, f and g.

PlotlyLightLite.plot! — Method

plot!([p::Plot], x, y; kwargs...)
plot!([p::Plot], f, a, [b]; kwargs...)
plot!([p::Plot], f; kwargs...)

Used to add a new tract to an existing plot. Like Plots.plot!. See plot for argument details.

PlotlyLightLite.plot — Method

plot(x, y, [z]; [linecolor], [linewidth], [legend], kwargs...)
plot(f::Function, a, [b]; kwargs...)

Create a line plot.

Returns a Plot instance from PlotlyLight

x,y points to plot. NaN values in y break the line

a, b: the interval to plot a function over can be given by two numbers or if just a then by extrema(a).

linecolor: color of line

linewidth: width of line

label in legend

Other keyword arguments include width and height, xlims and ylims, legend, aspect_ratio.

Provides an interface like Plots.plot for plotting a function f using PlotlyLight. This just scratches the surface, but PlotlyLight allows full access to the underlying JavaScript library library.

The provided "Plots-like" functions are plot, plot!, scatter!, scatter, annotate!, title!, xlims! and ylims!.

Example

p = plot(sin, 0, 2pi; legend=false)
plot!(cos)
# add points
x0 = [pi/4, 5pi/4]
scatter!(x0, sin.(x0), markersize=10)
# add text
annotate!(tuple(zip(x0, sin.(x0), ("A", "B"))...), halign="left", pointsize=12)
title!("Sine and cosine and where they intersect in [0,2π]")
# adjust limits
ylims!((-3/2, 3/2))
# add shape
y0, y1 = extrema(p).y
[rect!(xᵢ-0.1, xᵢ+0.1, y0, y1, fillcolor="gray", opacity=0.2) for xᵢ ∈ x0]
# display plot
p

Warning

You may need to run the first plot cell twice to see an image.

PlotlyLightLite.quiver! — Function

quiver!([p::Plot], x, y, txt=nothing; quiver=(dx, dy), kwargs...)
quiver(x, y, txt=nothing; quiver=(dx, dy), kwargs...)

(x,y) are tail positions, optionally labeled by txt

quiver specifies vector part of arrow

kwargs processarrowhead::Int?,arrowwidth::Int?,arrowcolor`

Example

ts = range(0, 2pi, length=100)
p = plot(sin.(ts), cos.(ts), linecolor="red")
ts = range(0, 2pi, length=10)
quiver!(p, cos.(ts), sin.(ts), quiver=(-sin.(ts), cos.(ts)), arrowcolor="red")
p

This example shows how text can be rotated with angles in degrees and positive angles measured in a clockwise direction.

ts = range(0, 2pi, 100)
p = plot(cos.(ts), sin.(ts), linecolor="red", aspect_ratio=:equal,
    linewidth=20, opacity=0.2)

txt = split("The quick brown fox jumped over the lazy dog")
ts = range(0, 360, length(txt)+1)[2:end]
for (i,t) ∈ enumerate(reverse(ts))
    quiver!(p, [cosd(t)],[sind(t)],txt[i],
            quiver=([0],[0]),
            textangle=90-t,
            font=(size=20,))
end
xaxis!(zeroline=false); yaxis!(zeroline=false) # remove zerolines
p

PlotlyLightLite.rect! — Method

rect!([p::Plot], x0, x1, y0, y1; kwargs...)

Draw rectangle shape.

Example

Use named tuple for line for boundary.

rect!(p, 2,3,-1,1; line=(color=:gray,), fillcolor=:red, opacity=0.2)

PlotlyLightLite.scatter! — Method

scatter(x, y, [z]; [markershape], [markercolor], [markersize], kwargs...)
scatter!([p::Plot], x, y, [z]; kwargs...)

Place point on a plot.

markershape: shape, e.g. "diamond" or "diamond-open"

markercolor: color e.g. "red"

markersize: size, as an integer

PlotlyLightLite.scatter — Method

scatter(x, y; kwargs...) see scatter!

PlotlyLightLite.scroll_zoom! — Method

scrollzoom!([p], x::Bool) turn on/off scrolling to zoom

PlotlyLightLite.size! — Method

size!([p::Plot]; [width], [height]) specify size of plot figure

PlotlyLightLite.surface — Method

surface(x, y, z; kwargs...)
surface!(x, y, z; kwargs...)
surface(x, y, f::Function; kwargs...)
surface!(x, y, f::Function; kwargs...)

Create surface plot.

PlotlyLightLite.title! — Method

title!([p::Plot], txt)

Set plot title.

PlotlyLightLite.unzip — Method

unzip(v, [vs...])
unzip(f::Function, a, b)

Take container of points, return vector of corrodinated. Reverse of zip. Uses code from the invert function of SplitApplyCombine.

The function version applies f to a range of points over (a,b) and then calls unzip. This uses the adapted_grid function from PlotUtils.

This uses

PlotlyLightLite.xaxis! — Method

xticks!([p::Plot]; kwargs...)
yticks!([p::Plot]; kwargs...)

Adjust ticks on chart.

ticks: a container or range

ticklabels: optional labels (same length as ticks)

showticklabels::Bool

PlotlyLightLite.xlims! — Method

xlims!(p, lims) set x limits of plot

PlotlyLightLite.ylims! — Method

ylims!(p, lims) set y limits of plot